% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_input}
\alias{.check_input}
\title{.check_input}
\usage{
.check_input(img, f_call = NULL)
}
\arguments{
\item{img}{An object that needs to checked.}

\item{f_call}{The name of the function inside which the
\code{.check_input} is called.}
}
\value{
An error message if the check fails.
}
\description{
\code{.check_input} is a helper function of the
\code{rquantae} package that checks whether the input is
a matrix of numeric or integer values. Error message are
thrown if that is not the case.
}
\keyword{internal}
