% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{.compl}
\alias{.compl}
\title{.compl}
\usage{
.compl(img, algorithm, rotate)
}
\arguments{
\item{img}{An array or matrix of numeric values or integer values, or the file path to the image.}

\item{algorithm}{character. Which compression algorithm to use.}

\item{rotate}{logical. Image rotation by 90 degrees?}
}
\value{
a numeric value (ratio compressed/uncompressed file size).
}
\description{
Returns the complexity of an image array / matrix or path.
}
\keyword{internal}
