% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{.compress_and_get_size}
\alias{.compress_and_get_size}
\title{.compress_and_get_size}
\usage{
.compress_and_get_size(img, flname, algorithm, use_existing_bmp)
}
\arguments{
\item{img}{An array or matrix of numeric values or integer values.}

\item{flname}{Character string (file name without extension).}

\item{algorithm}{Character string (which compression algorithm to use).}

\item{use_existing_bmp}{Logical (use existing bmp file).}
}
\value{
a numeric value (compressed file size)
}
\description{
Compresses an image and returns the compressed file size.
}
\keyword{internal}
