% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{.contr}
\alias{.contr}
\title{.contr}
\usage{
.contr(img)
}
\arguments{
\item{img}{A matrix of numeric values or integer values.}
}
\value{
a numeric value (RMS contrast)
}
\description{
Returns the RMS contrast of an image matrix.
}
\keyword{internal}
