% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-similarity.R
\name{.selfsim}
\alias{.selfsim}
\title{.selfsim}
\usage{
.selfsim(img, full, raw, logplot)
}
\arguments{
\item{img}{A matrix of numeric values or integer values, preferably
by square size.}

\item{full}{logical. Should the full frequency range be used for
interpolation?}

\item{raw}{logical. Should the raw value of the regression slope be returned?}

\item{logplot}{logical. Should the log-log power spectrum of the image be
plotted?}
}
\value{
a numeric value (self-similarity)
}
\description{
Returns the self-similarity of an image matrix as the degree to which the slope of the log-log power spectrum falls with power according to the value of 2.
}
\keyword{internal}
