% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\name{.shift_axis}
\alias{.shift_axis}
\title{.shift_axis}
\usage{
.shift_axis(img, imgW, xrange)
}
\arguments{
\item{img}{A matrix of numeric values or integer values.}

\item{imgW}{Image width (numeric).}

\item{xrange}{Shift range (numeric vector).}
}
\value{
a vector of mirror symmetry values
}
\description{
.shift_axis shifts the mirror axis by xrange and returns
the symmetries at each axis position by calling the
.sym_mirror function.
}
\keyword{internal}
