% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\name{.sym}
\alias{.sym}
\title{.sym}
\usage{
.sym(img, color, per_channel = TRUE, shift_range = 0.05)
}
\arguments{
\item{img}{A matrix of numeric values or integer values.}

\item{color}{logical. Color image?}

\item{per_channel}{logical. Channel-wise maximum symmetry?}

\item{shift_range}{numeric. Percentage of axis shift.}
}
\value{
one numeric value as the symmetry
}
\description{
Calculates the mirror symmetry of an
image by correlating image halves.
}
\keyword{internal}
