% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\name{.sym_mirror}
\alias{.sym_mirror}
\title{.sym_mirror}
\usage{
.sym_mirror(img)
}
\arguments{
\item{img}{A matrix of numeric values or integer values.}
}
\value{
a numeric value between 0 and 1
}
\description{
sym_mirror returns the mirror symmetry of a grayscale
image matrix. To this end, the left and right image
halves are correlated.
}
\keyword{internal}
