% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typicality.R
\name{.typ}
\alias{.typ}
\title{.typ}
\usage{
.typ(imglist)
}
\arguments{
\item{imglist}{A list of matrices with numeric values or
integer values.}
}
\value{
a numeric value (RMS contrast)
}
\description{
Returns the typicality of a list of images as the correlation with the mean image.
}
\keyword{internal}
