% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_berries}
\alias{img_berries}
\title{Image of berries}
\format{A 3-dimensional array with 673 rows and 1000 columns each.}
\source{
\url{http://getrefe.com/downloads/berries-2/}
}
\usage{
img_berries
}
\description{
An array representing a colored image of berries.
}
\keyword{datasets}

