% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_bike}
\alias{img_bike}
\title{Image of a bike against a yellow wall}
\format{A 3-dimensional array with 750 rows and 1000 columns each.}
\source{
\url{http://getrefe.com/downloads/yellow-wall/}
}
\usage{
img_bike
}
\description{
An array representing a colored image of a bike leaning against a yellow wall.
}
\keyword{datasets}

