% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_bridge}
\alias{img_bridge}
\title{Close up image of the Golden Gate Bridge}
\format{A 3-dimensional array with 667 rows and 1000 columns each.}
\source{
\url{http://publicdomainarchive.com/public-domain-images-golden-gate-bridge-orange-blue-sky-cables/}
}
\usage{
img_bridge
}
\description{
An array representing a colored close up image of the Golden Gate Bridge.
}
\keyword{datasets}

