% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_fireworks}
\alias{img_fireworks}
\title{Image of fireworks}
\format{A 3-dimensional array with 662 rows and 1000 columns each.}
\source{
\url{http://getrefe.com/downloads/fireworks-2/}
}
\usage{
img_fireworks
}
\description{
An array representing a colored image of fireworks.
}
\keyword{datasets}

