% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_office}
\alias{img_office}
\title{Image of an office desk}
\format{A 3-dimensional array with 750 rows and 1000 columns each.}
\source{
\url{http://publicdomainarchive.com/public-domain-images-letter-press-screen-printing-office-desk-work-t-shirt-camera-wood/}
}
\usage{
img_office
}
\description{
An array representing a colored image a office desk with letter press utilities.
}
\keyword{datasets}

