% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_rails}
\alias{img_rails}
\title{Image of rails}
\format{A 3-dimensional array with 667 rows and 1000 columns each.}
\source{
\url{http://getrefe.com/downloads/man/}
}
\usage{
img_rails
}
\description{
An array representing a colored image of a man walking on rails.
}
\keyword{datasets}

