% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_romanesco}
\alias{img_romanesco}
\title{Image of romanesco}
\format{A 3-dimensional array with 734 rows and 1000 columns each.}
\source{
\url{https://commons.wikimedia.org/wiki/File:Romanesco_Brassica_oleracea_Richard_Bartz.jpg}
}
\usage{
img_romanesco
}
\description{
An array representing a colored image of romanesco broccoli.
}
\keyword{datasets}

