% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_sky}
\alias{img_sky}
\title{Image of the sky}
\format{A 3-dimensional array with 667 rows and 1000 columns each.}
\source{
\url{http://publicdomainarchive.com/public-domain-images-clouds-sky-mountain-blue/}
}
\usage{
img_sky
}
\description{
An array representing a colored image of the sky above the clouds.
}
\keyword{datasets}

