% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_trees}
\alias{img_trees}
\title{Image of trees}
\format{A 3-dimensional array with 667 rows and 1000 columns each.}
\source{
\url{http://publicdomainarchive.com/public-domain-images-tall-forest-green-trees/}
}
\usage{
img_trees
}
\description{
An array representing a colored image of trees in a forest.
}
\keyword{datasets}

