% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_valley_green}
\alias{img_valley_green}
\title{Image of a green valley}
\format{A 3-dimensional array with 662 rows and 1000 columns each.}
\source{
\url{http://publicdomainarchive.com/public-domain-images-mountains-green-river-valley-forest-nature/}
}
\usage{
img_valley_green
}
\description{
An array representing a colored image of a green valley.
}
\keyword{datasets}

