% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{img_valley_white}
\alias{img_valley_white}
\title{Image of a snowy valley}
\format{A 3-dimensional array with 662 rows and 1000 columns each.}
\source{
\url{http://publicdomainarchive.com/public-domain-images-mountain-snow-peak-blue-sky/}
}
\usage{
img_valley_white
}
\description{
An array representing a colored image of a snow-covered valley.
}
\keyword{datasets}

