% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{quantify_contrast}
\alias{quantify_contrast}
\title{RMS Contrast of an Image}
\usage{
quantify_contrast(img)
}
\arguments{
\item{img}{A matrix of numeric values or integer values.
Color images have to be converted to grayscale in
advance (function \code{rgb2gray}) or each color
channel has to be analyzed seperately.}
}
\value{
a numeric value (RMS contrast)
}
\description{
\code{quantify_contrast} returns the RMS contrast of an
image matrix \code{img}. A higher value indicates higher
contrast.
}
\details{
The function returns the RMS contrast of an
  image matrix \code{img}. The RMS contrast is defined as
  the standard deviation of the normalized pixel
  intensity values. A higher value indicates higher
  contrast. The image is automatically normalized if
  necessary (i.e., normalization into range [0, 1]).
}
\examples{
# Example image with high contrast: img_berries
#
# display image
grid::grid.raster(img_berries)
# convert to grayscale
berries_grayscale <- rgb2gray(img_berries)
# get contrast
quantify_contrast(berries_grayscale)

# Example image with low contrast: img_bike
#
# display image
grid::grid.raster(img_bike)
# convert to grayscale
bike_grayscale <- rgb2gray(img_bike)
# get contrast
quantify_contrast(bike_grayscale)

}
\references{
Peli, E. (1990). Contrast in complex images.
  \emph{Journal of the Optical Society of America A},
  \emph{7}, 2032--2040.
  doi:\href{https://doi.org/10.1364/JOSAA.7.002032}{10.1364/JOSAA.7.002032}
}
\seealso{
\code{\link{rgb2gray}},
  \code{\link{quantify_symmetry}},
  \code{\link{quantify_complexity}},
  \code{\link{quantify_typicality}},
  \code{\link{quantify_self_similarity}}
}

