% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\name{quantify_symmetry}
\alias{quantify_symmetry}
\title{quantify_symmetry}
\usage{
quantify_symmetry(img)
}
\arguments{
\item{img}{A matrix of numeric values or integer values.
Color images have to be converted to grayscale in
advance or each color channel has to be analyzed
seperately.}
}
\value{
a list of numeric values
}
\description{
quantify_symmetry returns the vertical and horizontal
symmetry of an image. Values can range between 0 (not
symmmetrical) and 1 (fully symmetrical).
}
\examples{
# construct sample image
img <- matrix(0, nrow=100, ncol=100)
img[21:60,31:70] <- 0.5
img[71:90, 21:80] <- 1

# show image
OpenImageR::imageShow(img)

# get both vertical and horizontal symmetry
quantify_symmetry(img)
}

