% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typicality.R
\name{quantify_typicality}
\alias{quantify_typicality}
\title{Typicality of Images relative to each other}
\usage{
quantify_typicality(imglist, rescale = NULL)
}
\arguments{
\item{imglist}{A list of matrices with numeric values or
integer values. Color images have to be converted to
grayscale in advance (function \code{rgb2gray}) or each
color channel has to be analyzed seperately.}

\item{rescale}{numeric. Rescales the images prior to
computing the typicality scores (per default no
rescaling is performed). Rescaling is performed by the
\code{resizeImage} function from the \code{OpenImageR}
package (bilinear rescaling)}
}
\value{
a list with a matrix of numeric value (typicality
  scores)
}
\description{
\code{quantify_typicality} returns the visual typicality
of a list of image matrices \code{imglist} relative to
each other. Higher values indicate larger typicality.
}
\details{
The function returns the visual typicality of a
  list of image matrices \code{imglist} relative to each
  other, while higher values are indicating larger
  typicality.

  The typicality score is computed as the correlation of
  a particular image with the average representation of
  all images, i.e. the mean of all images.

  Rescaling of the images prior to computing the
  typicality scores can be specified with the optional
  rescaling parameter (must be a numeric value).
}
\examples{
# construct a list of 20 sample images
imgs <- replicate(20,
  matrix(runif(100, min = 0, max = 255), nrow = 10, ncol = 10),
  simplify = FALSE)

# get typicality scores
quantify_typicality(imgs)

}
\references{
Mayer, S. & Landwehr, J. R. (2016). Measuring
  design typicality -- a comparison of objective and
  subjective approaches. \emph{Proceedings of DRS 2016,
  Design Research Society 50th Anniversary Conference}.
  Brighton, UK, 27--30 June 2016.
}
\seealso{
\code{\link{rgb2gray}},
  \code{\link{quantify_symmetry}},
  \code{\link{quantify_complexity}},
  \code{\link{quantify_contrast}},
  \code{\link{quantify_self_similarity}}
}

