% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\name{sym_hor}
\alias{sym_hor}
\title{sym_hor}
\usage{
sym_hor(img)
}
\arguments{
\item{img}{}
}
\value{
a numeric value between 0 and 1
}
\description{
sym_hor returns the horizontal symmetry of a grayscale
image matrix. To this end, the upper and lower image
halves are correlated.
}

