% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{quantify_contrast}
\alias{quantify_contrast}
\title{RMS Contrast of an Image}
\usage{
quantify_contrast(img, normalize = TRUE)
}
\arguments{
\item{img}{A matrix of numeric values or integer values.
Color images have to be converted to grayscale in
advance (function \code{rgb2gray}) or each color
channel has to be analyzed seperately. The image is
assumed to have its pixel intensities \strong{not}
normalized but to be in the range [0, 255]}

\item{normalize}{logical. Should pixel intensity
normalization into range [0, 1] be performed?}
}
\value{
a numeric value (RMS contrast)
}
\description{
\code{quantify_contrast} returns the RMS contrast of an
image matrix \code{img}. A higher value indicates higher
contrast.
}
\details{
The function returns the RMS contrast of an
  image matrix \code{img}. The RMS contrast is defined as
  the standard deviation of the normalized pixel
  intensity values. A higher value indicates higher
  contrast.

  As the function assumes that the pixel intensity values
  of the image \code{img} are in the range [0, 255],
  pixel normalization into [0, 1] is done by default. If
  \code{normalize} is set to \code{FALSE} no
  normalization is performed.
}
\examples{
# construct sample image
img <- matrix(0, nrow=100, ncol=100)
img[21:80, 21:80] <- 255

# if you want to inspect the image
# OpenImageR::imageShow(img)

# get contrast
quantify_contrast(img)

}
\references{
Peli, E. (1990). Contrast in complex images.
  \emph{Journal of the Optical Society of America A},
  \emph{7}, 2032--2040.
  doi:\href{https://doi.org/10.1364/JOSAA.7.002032}{10.1364/JOSAA.7.002032}
}
\seealso{
\code{\link{rgb2gray}},
  \code{\link{quantify_symmetry}},
  \code{\link{quantify_complexity}},
  \code{\link{quantify_typicality}},
  \code{\link{quantify_self_similarity}}
}

