% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\name{quantify_symmetry}
\alias{quantify_symmetry}
\title{Symmetry of an Image}
\usage{
quantify_symmetry(img, vertical = TRUE, horizontal = TRUE)
}
\arguments{
\item{img}{A matrix of numeric values or integer values.
Color images have to be converted to grayscale in
advance (function \code{rgb2gray}) or each color
channel has to be analyzed seperately.}

\item{vertical}{logical. Should the vertical symmetry be
computed?}

\item{horizontal}{logical. Should the horizontal symmetry
be computed?}
}
\value{
a named vector of numeric values (vertical and horizontal
  symmetry)
}
\description{
\code{quantify_symmetry} returns the vertical and
horizontal symmetry of an image matrix \code{img}.
}
\details{
The function returns the vertical and horizontal
  symmetry of an image matrix \code{img}. Values can
  range between 0 (not symmmetrical) and 1 (fully
  symmetrical). If \code{vertical} or \code{horizontal}
  is set to \code{FALSE} then vertical or horizontal
  symmetry is not computed, respectively.
}
\examples{
# Example image with high vertical symmetry: img_rails
#
# display image
grid::grid.raster(img_rails)
# convert to grayscale
rails_grayscale <- rgb2gray(img_rails)
# get symmetry
quantify_symmetry(rails_grayscale)

# Example image with low vertical symmetry: img_bridge
#
# display image
grid::grid.raster(img_bridge)
# convert to grayscale
bridge_grayscale <- rgb2gray(img_bridge)
# get symmetry
quantify_symmetry(bridge_grayscale)

}
\references{
Mayer, S. & Landwehr, J. R. (2014). When
  Complexity is Symmetric: The Interplay of Two Core
  Determinants of Visual Aesthetics. \emph{Advances in
  Consumer Research}, \emph{42}, 608--609.
}
\seealso{
\code{\link{rgb2gray}},
  \code{\link{quantify_complexity}},
  \code{\link{quantify_contrast}},
  \code{\link{quantify_typicality}},
  \code{\link{quantify_self_similarity}}
}

