% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\name{quantify_symmetry}
\alias{quantify_symmetry}
\title{Symmetry of an Image}
\usage{
quantify_symmetry(img, vertical = TRUE, horizontal = TRUE)
}
\arguments{
\item{img}{A matrix of numeric values or integer values.
Color images have to be converted to grayscale in
advance or each color channel has to be analyzed
seperately.}

\item{vertical}{logical. Should the vertical symmetry be
computed?}

\item{horizontal}{logical. Should the horizontal symmetry
be computed?}
}
\value{
a list of numeric values (vertical and horizontal
  symmetry)
}
\description{
\code{quantify_symmetry} returns the vertical and
horizontal symmetry of an image matrix \code{img}. Values
can range between 0 (not symmmetrical) and 1 (fully
symmetrical). If \code{vertical} or \code{horizontal} is
set to \code{FALSE} then vertical or horizontal symmetry
is not computed, respectively.
}
\examples{
# construct sample image
img <- matrix(0, nrow=100, ncol=100)
img[21:60,31:70] <- 0.5
img[71:90, 21:80] <- 1

# if you want to inspect the image
# OpenImageR::imageShow(img)

# get both vertical and horizontal symmetry
quantify_symmetry(img)

}
\references{
Mayer, S. & Landwehr, J. R. (2014). When
  Complexity is Symmetric: The Interplay of Two Core
  Determinants of Visual Aesthetics. \emph{Advances in
  Consumer Research}, \emph{42}, 608--609.
}

