% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagefluencyApp.R
\name{run_imagefluency}
\alias{run_imagefluency}
\title{Run imagefluency app}
\usage{
run_imagefluency()
}
\description{
Launches a Shiny app that shows a demo of what can be done with
the \code{imagefluency} package.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  run_imagefluency()
}
}
